feat(documentation): project description
- Added requirements to start the service
- Information about environment variables
- Commands are described

+++ pkg/configing/config.go
@@ -123,7 +123,7 @@ type ClientConfig struct {
 	Endpoint      string
 	ReadTimeout   int
 	WriteTimeout  int
-	UpdateMinutes int
+	UpdateMinutes float64
 }
 
 // NewClientConfig constructor for ClientConfig
@@ -133,7 +133,7 @@ func NewClientConfig() *ClientConfig {
 		Endpoint:      cfg.GetString(httpClientEndpoint),
 		ReadTimeout:   cfg.GetInt(httpClientReadTimeout),
 		WriteTimeout:  cfg.GetInt(httpClientWriteTimeout),
-		UpdateMinutes: cfg.GetInt(httpClientUpdateMinutes),
+		UpdateMinutes: cfg.GetFloat64(httpClientUpdateMinutes),
 	}
 }
